Move useSubDashboards to a separated file
