+++ packages/lesmis/src/components/exports/DashboardExportModal.js
@@ -1,34 +1,16 @@
 /*
  * Tupaia
- *  Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
+ *  Copyright (c) 2017 - 2022 Beyond Essential Systems Pty Ltd
  */
-import React, { useMemo, useState } from 'react';
+import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import DownloadIcon from '@material-ui/icons/GetApp';
 import Pagination from '@material-ui/lab/Pagination';
 import { Dialog, DialogHeader, DialogContent, FlexCenter } from '@tupaia/ui-components';
 import MuiIconButton from '@material-ui/core/IconButton';
 import { DashboardExportPreview } from './DashboardExportPreview';
-import { useDashboardDropdownOptions } from '../../utils/useDashboardDropdownOptions';
-import { exportImagesToPDF, useGetImages, useUrlParams } from '../../utils';
-import { useDashboardData } from '../../api';
-
-const useSubDashboards = () => {
-  const { entityCode } = useUrlParams();
-  const { dropdownOptions } = useDashboardDropdownOptions();
-  const profileDropDownOptions = dropdownOptions.find(({ value }) => value === 'profile');
-  const { filterSubDashboards } = profileDropDownOptions.componentProps;
-  const { data } = useDashboardData({
-    entityCode,
-    includeDrillDowns: false,
-  });
-  const subDashboards = useMemo(() => data?.filter(filterSubDashboards), [
-    data,
-    filterSubDashboards,
-  ]);
-
-  return subDashboards;
-};
+import { exportImagesToPDF, useGetImages } from '../../utils';
+import { useSubDashboards } from './utils/useSubDashboards';
 
 export const DashboardExportModal = ({ Button, title }) => {
   const [isOpen, setIsOpen] = useState(false);
