+++ packages/lesmis/src/components/exports/utils/index.js
@@ -1,0 +1,1 @@
+export { useSubDashboards } from './useSubDashboards';
